v4l2_codec2: Add option to e2e encoder test to use software encoder.
This CL adds a "--use_sw_encoder" parameter to the e2e encoder test,
that makes the test use a SW encoder instead of the default HW encoder.
This can be used to verify functioning of the software encoder, and to
provide a performance baseline for the HW video encoder tests.

Bug: 169626661
Test: tast run DUT arc.VideoEncodeAccel.h264_1080p_i420 on hatch
Change-Id: I0289aa5e569bd7229ba18ad57b5f2e0909d60f1c