v4l2_codec2: Add new SW decoder names to the e2e decoder test.
This CL adds the new names used for the SW decoders on Android R to the
e2e decoder test. The old names are currently still supported as an
alias of the new decoder names, but we should switch to using the new
names to make the tests future-proof.

New SW encoder names:
- OMX.google.h264.decoder -> c2.android.avc.decoder
- OMX.google.vp8.decoder -> c2.android.vp8.decoder
- OMX.google.vp9.decoder -> c2.android.vp9.decoder

Bug: None
Test: tast run DUT arc.VideoDecodeAccel.h264 on hatch
Change-Id: Ie84fcbe4f93943e7bc6be4d2d95163c369c06047