v4l2_codec2: Split up V4L2EncodeComponent.
This CL splits the V4L2EncodeComponent in a codec2 component and a V4L2
encoder part. This improves the encapsulation and flexibility of the
encoder and makes maintenance easier.

As a lot of the encode/drain/flush logic has been affected by the
split, various changes needed to be made:
- The drain/flush flows have been reworked.
- Various changes have been made to the input/output buffer flows to
  match the new requirements and restrictions.
- The BitstreamBuffer class is used for output, to avoid having Codec2
  logic in the V4L2 encoder class.
- A queue was introduced for work waiting for input format conversion.
- Some outdated comments have been removed.

Bug: 143333813
Test: arc.VideoEncodeAcel.* and various CTS tests
Change-Id: Ie782b74daa3662a1ce1d441b5ba5e48eaa1057d3