C2VdaBqBlockPool: remove fence timeout
Remove timeout when waiting on the fence of dequeued buffers. Fences are
guaranteed to fire within a relatively short period of time. Since the
wait is already performed on a dedicated thread, waiting until the fence
fires is safe. Omitting the timeout reduces the number of IPCs
performed, as buffers are no longer cancelled and re-dequeued.

Bug: 172184040
Test: tast run DUT arc.VideoDecodeAccel.vp8_vm

Change-Id: I7c7f444c340f75575fd985416a29bbfeafdaf9e6