components: Set InputBufferManager interval
InputBufferManager notifies the codec2 framework when the component no
longer has any references to an input buffer. While a dedicated release
notification is necessary in some circumstances, the release can usually
be done when the corresponding work item is completed, to reduce the
amount of IPC. This change sets the notification interval used by
InputBufferManager to reduce how often a dedicated release notification
is sent.

Bug: 172184040
Test: tast run DUT arc.VideoDecodeAccel.vp8_vm

Change-Id: Id6371d328bb00730b50d3b3075eccf4010554f27