C2VdaBqBlockPool: Rewrite configureProducer
C2VdaBqBlockPool::configureProducer() is used to switch surface.
Originally the flow of changing surface is for C2VDAComponent.
This CL rewrites the flow of configureProducer for
V4L2DecodeComponent/V4L2Decoder/VideoFramePool structure.

1. During surface switching, we restrict the maximum dequeued buffer
count to the count of the current tracking buffers. Because the codec2
framework would migrate some buffers to the new surface. restricting
the dequeued buffer count can guarantee any buffer is attached to
multiple slots.

2. Due to 1., there might be not enough slots to migrate all the
buffers at the time of switching surface. We defer the migration
process to fetchGraphicBlock(), because the client would call this
method when a slot is released.

3. After all buffers are attached to the new surface, these attached
buffers are considered dequeued. We store the slots indices in
|mDequeuedSlots| and return them to fetchGraphicBlock() from
getFreeSlotLocked().

Bug: 162902320
Test: android.media.cts.MediaCodecPlayerTest#testPlaybackSwitchViews
Test: android.media.cts.AdaptivePlaybackTest#testH264_*
Change-Id: I457d3b5cb44afcfa022ba4902570126b81a06c4f