v4l2_codec2: Configure correct stride when encoding.
When allocating input buffers the gralloc allocator talks to minigbm
which uses a stride of 64. This causes QBUF failures when queuing
buffers on the input queue as the stride does not match with what the
virtio video driver expects.

This CL makes changes to communicate the actual stride to the virtio
encoder using v4l2_pix_format.bytesperline.

Bug: 173083834
Bug: 173666696
Test: android.video.cts.VideoEncoderDecoderTest.testAvcOther0Perf0720x0480

Change-Id: I798d0bc694f5d52f19fb7056523baccdf04f91a2