v4l2_codec2: pass SW flags in getGraphicBlockInfo
On ARCVM, using no SW flags when locking a buffer causes it to return
the host buffer stride/offsets. This isn't what is desirable here, as
the buffer dimensions here are used for validation of the guest buffer
within the guest kernel. To get the guest stride/offsets, we need to
pass SW flags when locking the buffer.

Note that crosvm switches to using the host stride/offsets before
passing the buffers to libvda.

Bug: 173667442
Test: android.media.cts.VideoEncoderTest#testOtherH264SurfArbitraryW

Change-Id: I1c87a12179fd3ed3f17c0414f7266506746a96b9