C2VdaBqBlockPool: Polish the retry logic
After fetching new IGBP slot times out, we should retry the fetching
when buffer released at IGBP. We use a flag
|mBufferReleasedAfterTimedOut| to catch the case that the buffer is
released between fetching timed out and setting the retry callback.
If the flag is set, then we should execute the callback immediately,
instead of waiting for the next buffer released event.

However, we forgot to reset the flag when executing the callback. It
caused we continuously retry the buffer fetching until the next buffer
released event. This CL fixes this issue.

Bug: 175489166
Test: run com.google.android.exoplayer.gts.DashTest#testH264Adaptive
      and check setMaxDequeuedBufferCount() only retry when buffer
      released
Change-Id: I8f7a36af50e5eebfcbecedc0f126c86f9427fa95