V4L2DecodeComponent: store weak_ptr at mDecoderTaskRunner
The destructor joins mDecoderThread to prevent use-after-free errors.
However, some of the tasks that mDecoderTaskRunner executes try to
call V4L2DecodeComponent::shared_from_this, which can fail because
creating a shared pointer once an object's destructor has already
started running is not safe.

This CL fixes this issue by storing a weak_ptr at mDecoderTaskRunner.
We lock the weak_ptr to get the shared_ptr when the task on
mDecoderTaskRunner needs.

Bug: 172025534
Test: android.media.cts.AdaptivePlaybackTest#testH264_*
Change-Id: Ib35efbdd8e47cbee6455b5673bb8303384fbb086