C2VdaBqBlockPool: Use GEM handle as the identifier of graphic buffers.
Originally we assume the buffers at BufferQueue are not moved to
different slot. Then we could track the buffers by the slot index.
However, MediaSync implementation detach buffers and attach back to
BufferQueue. The slot index might be changed.

This CL changes to use GEM handle as the identifier of graphic
buffers. Because all the graphic buffers are allocated by gralloc, we
could get the GEM handle by DRM. V4L2GraphicAllocator is not used
anymore.

Bug: 177484971
Test: android.media.cts.AdaptivePlaybackTest

Change-Id: I6b7f3f9a6c1d6b97d0e44aa2a19bc71ad3bab880