C2VdaBqBlockPool: track graphic buffer by identifier
Originally we assumed the buffers at BufferQueue is not moved to
different slot, so we tracked the graphic buffers by storing tuple of
(slot index, unique ID, C2Allocation). However, the slot index of the
buffer might be changed.

In this CL, we change to store two maps instead:
- unique ID to C2Allocation
- slot index to GraphicBuffer

The map from unique ID to C2Allocation is for tracking the fixed
buffer set. The map from slot index to GraphicBuffer is updated when
IGBP::dequeueBuffer() returns BUFFER_NEEDS_REALLOCATION, no matter if
the buffers is tracked or not. With these two maps, we can handle the
case that the consumer move graphic buffers to different BufferQueue
slot.

Bug: 177484971
Test: android.media.cts.AdaptivePlaybackTest
Test: android.media.cts.MediaSyncTest

Change-Id: Iee99e8d0b00dae463cf6483ca8c971cc84f9ea49