C2VdaBqBlockPool: remove unnecessary data
|mComponentOwnedUniquedIds| was used to prevent the issue that buffers
might be attached twice to new surface. But ag/13490974 change the way
to switch surface. This issue doesn't exist anymore. This CL removes
the unnecessary mComponentOwnedUniquedIds data.

Also, because the extra field |mUniqueId| of C2VdaBqBlockPoolData is
used to update |mComponentOwnedUniquedIds|, it's not needed as well.
This CL removes C2VdaBqBlockPoolData class and uses its parent class
directly.

Bug: 174188958
Test: android.media.cts.AdaptivePlaybackTest
Test: android.media.cts.MediaCodecPlayerTest#testPlaybackSwitchViews

Change-Id: I588f3896a8ad6cceb90c844981e7ab95defe2092