v4l2_codec2: Expand NALParser to support parsing color aspects.
This CL expands the simple NALParser to support parsing color aspects
from the SPS NAL unit. The V4L2 decode component is adapted to use this
new NALParser rather than the Chrome parser. This allows us to remove
the Chrome H.264 parser and its associated dependencies in a subsequent
CL, removing a lot of code.

Bug: 155138142
Test: arc.VideoDecodeAccel.* and arc.VideoDecodeAccelPerf.* on hatch
Change-Id: Icbfa63478980ab2b2d1bf2d46697359a0ac92937