v4l2_codec2: Merge VideoFrame into VideoPixelFormat.
After the recent cleanup operations the only code left in the
VideoFrame class are helper functions directly related to the
VideoPixelFormat. This CL merges all these helpers into
video_pixel_format.h.

Additionally a helper in the VideoFrameLayout class is also moved to
the video_pixel_format.h, so all VideoPixelFormat-related helpers are
in one place.

Bug: 155138142
Test: arc.VideoEncodeAccel.h264_192p_i420_vm
Change-Id: I76699ecf69eb90fe890c1c272db058938b68062b