v4l2_codec2: Move media::VideoPixelFormat to common code.
This CL Moves the VideoPixelFormat files to the common code folder. In a
next step these files will be cleaned up and reworked into proper
Android code, but splitting the move into a separate CL makes the review
process a lot easier.

Note: the "accel" folder is now empty, but the folder and Android.bp
      files will be cleaned up in a next CL so we can properly clean
      up all dependencies on the v4l2_accel module.

Bug: 155138142
Test: arc.VideoEncodeAccel.h264_192p_i420_vm
Change-Id: Ia4708e5dbdd11223b9a2c445850b8c0cf6ee6d37