v4l2_codec2: Remove minimum H.264 profile workaround in V4L2 encoder.
This CL removes a workaround that was added to tackle a quality issue
in b/114332827. This workaround causes the android.mediav2.cts.
EncoderProfileLevelTest#testValidateProfileLevel[6(video/avc)]
test to fail.

It's not clear whether the issue is still present so some quality check
CTS tests might start failing, but we should tackle these issues at the
source.

Bug: 184015100
Test: android.mediav2.cts.EncoderProfileLevelTest
Change-Id: I55f275af4a421509a776e39bed669365159ba2ee