v4l2_codec2: Use graceful methods to read NAL color aspects.
This CL changes the color aspects H.264 NAL parser to use
getBitsGraceful() when reading from the NALU, and check the result of
bit operations. Reading from the NALU is currenlty safe and the bit
reader never reads out-of-bounds, but an assert is thrown causing the
StagefrightTest to abort rather than gracefully fail.

Bug: 187402250
Test: arc.VideoDecodeAccel.* on hatch
Change-Id: I6abb5f485e0a6f4fb8c46c8b74e895127e9107b5