v4l2_codec2 encoder: Don't parse H.264 NAL units when using VP8/9.
This CL changes the V4L2 encode component to not try to extract H.264
NAL units from an encoded VP8/9 video stream. Parsing for NAL units in
a VP8/9 stream doesn't break encoding as an empty list is returned, but
returning an empty config update to the mediacodec framework makes the
GTS VP8 RtcVideoCodecTest fail.

BUG: 191644279
BUG: 155138243
Test: media.gts.RtcVideoCodecTest#testDynamicBitrateChangeVp8
Change-Id: I789eb0c948afaea29e1595a503d63cd3788baf1b