v4l2_codec2 encoder: Add support for prepending SPS/PPS to IDR frames.
The V4L2 encoder requests SPS and PPS NAL units to be prepended to IDR
frames using the V4L2_CID_MPEG_VIDEO_PREPEND_SPSPPS_TO_IDR control, to
improve the resilience of the encoded video stream. When this control
is not supported a warning is logged but no other steps are taken.

This CL adds functionality to manually cache and preprend SPS and PPS
NAL units to IDR frames when the above control is unsupported.

BUG: 161495502
Test: arc.VideoEncodeAccel.h264_192p_i420_vm

Change-Id: I6b0906c4bdd46041f34eab7061e61f7b4a460fe2